+++ src/idc/paddle/class_lister.py
@@ -4,6 +4,7 @@ from typing import List, Dict
 def list_classes() -> Dict[str, List[str]]:
     return {
         "seppl.io.Reader": [
+            "idc.paddle.reader",
             "idc.paddle.reader.imgcls",
             "idc.paddle.reader.imgseg",
         ],