+++ src/idc/paddle/reader/__init__.py
@@ -1,0 +1,1 @@
+from ._visualdl import VisualDLReader